move header component to layout folder
